+++ scripts/validate_pdf_fixture.sh
@@ -18,8 +18,10 @@ swiftc \
   "${ROOT_DIR}/Skald/Services/Converters/DocumentConverter.swift" \
   "${ROOT_DIR}/Skald/Models/OutputFormat.swift" \
   "${ROOT_DIR}/Skald/Models/ConversionError.swift" \
+  "${ROOT_DIR}/Skald/Models/DelimitedOptions.swift" \
   "${ROOT_DIR}/Skald/Models/SourceFileDescriptor.swift" \
   "${ROOT_DIR}/Skald/Models/ReadableModels.swift" \
+  "${ROOT_DIR}/Skald/Models/CanonicalDelimitedTable.swift" \
   "${ROOT_DIR}/Skald/Support/Formatting/ReadableOutputFormatter.swift" \
   "${ROOT_DIR}/Skald/Support/Parsing/PlainTextParser.swift" \
   "${ROOT_DIR}/Skald/Support/Parsing/AttributedTextParser.swift" \